Trainer: Move Statistics to Trainer.py
