+++ onmt/__init__.py
@@ -1,11 +1,10 @@
 import onmt.IO
 import onmt.Models
 import onmt.Loss
-from onmt.Trainer import Trainer
+from onmt.Trainer import Trainer, Statistics
 from onmt.Translator import Translator
 from onmt.Optim import Optim
 from onmt.Beam import Beam, GNMTGlobalScorer
-from onmt.Statistics import Statistics
 
 
 # For flake8 compatibility
